LayoutTests/fast/regions/selection/crash-deselect-expected.txt

CONSOLE MESSAGE: line 66: HierarchyRequestError (DOM Exception 3): The operation would yield an incorrect node tree.

Errlog selectionTest: HierarchyRequestError: The operation would yield an incorrect node tree. : index 1 
Errlog webtest_fn_24: 
